requirements.txt
openai
google-generativeai
transformers
torch
pandas
pillow
beautifulsoup4
python-dotenv
openpyxl
nltk
